iniciando a remodelagem para o padrao mvc

+++ controller/controllerCategory.py
@@ -1,5 +1,6 @@
 from model.category import Category
 from model.daoCategory import CategoryDao
+from exception.NotExistsException import NotExistsException
 
 
 class ControllerCategory:
@@ -14,28 +15,38 @@ class ControllerCategory:
         else:
             raise TypeError
 
-    def update(self, id: int, name=None):
+    def update(self, id: int, name: str):
         if isinstance(id, int):
             category = self.__dao.read(id)
-            if name:
-                if isinstance(name, str):
-                    category.name = name
-                else:
-                    raise TypeError
-            return self.__dao.update(category)
+            if category:
+                if name:
+                    if isinstance(name, str):
+                        category.name = name
+                    else:
+                        raise TypeError
+                return self.__dao.update(category)
+            else:
+                raise NotExistsException
         else:
             raise TypeError
 
     def delete(self, id: int):
         if isinstance(id, int):
             category = self.__dao.read(id)
-            return self.__dao.delete(category)
+            if category:
+                return self.__dao.delete(category)
+            else:
+                raise NotExistsException
         else:
             raise TypeError
 
     def read(self, id: int):
         if isinstance(id, int):
-            return self.__dao.read(id)
+            category = self.__dao.read(id)
+            if category:
+                return category
+            else:
+                raise NotExistsException
         else:
             raise TypeError
